Autoeconder is imported only if Keras is available.
