+++ hermione/core/analysis/__init__.py
@@ -1,5 +1,5 @@
 from ._cluster import Clusterer
-from ._dimensionality_reduction import Autoencoder, DimensionalityReducer
+from ._dimensionality_reduction import DimensionalityReducer
 from ._feature_selection import FeatureSelector
 from ._hypothesis_autopilot import HTestAutoPilot
 from ._hypothesis_testing import HypothesisTester
@@ -9,7 +9,6 @@ from ._vif import VIF
 
 __all__ = [
     'Clusterer',
-    'Autoencoder', 
     'DimensionalityReducer',
     'FeatureSelector',
     'HTestAutoPilot',
@@ -17,3 +16,10 @@ __all__ = [
     'PCAReducer',
     'VIF'
 ]
+
+
+try:
+    from ._autoenconder import Autoencoder
+    __all__.append('Autoencoder')
+except (ModuleNotFoundError, ImportError):
+    pass
